rename 'names' module 'transactors'
